Use .yml files for frontend GUI loc
